requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
numpy==1.24.3  # Downgrade to compatible version
opencv-python==4.8.1.78
python-multipart==0.0.6
scikit-learn==1.3.2
Pillow==10.0.1
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
joblib==1.3.2